change datatype of salary and enhance a bit the form view

+++ src/main/java/com/springassignment/demo/Model/Employee.java
@@ -2,6 +2,7 @@ package com.springassignment.demo.Model;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import java.math.BigDecimal;
 
 public class Employee {
 
@@ -9,13 +10,13 @@ public class Employee {
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
     private String name;
-    private float salary;
+    private BigDecimal salary;
 
     public Employee() {
 
     }
 
-    public Employee(Long id, String name, float salary) {
+    public Employee(Long id, String name, BigDecimal salary) {
         this.id = id;
         this.name = name;
         this.salary = salary;
@@ -37,11 +38,11 @@ public class Employee {
         this.name = name;
     }
 
-    public float getSalary() {
+    public BigDecimal getSalary() {
         return salary;
     }
 
-    public void setSalary(float salary) {
+    public void setSalary(BigDecimal salary) {
         this.salary = salary;
     }
 }
